move dummy data generator to utils
